Implement a basic Pythia event generation file
The file generate_events.py will generate a requested number of events
for a given mass and width for the W boson. Currently, I have to figure
out why the width is not changing. Additionally, I need to implement
a bash script to run over all the events to generate.

Maybe discuss if current format is acceptable.